Cleaning up Task Runner quickstart, and updating documentation
