+++ docs/requirements-docs.txt
@@ -3,6 +3,7 @@
 sphinx-rtd-theme
 sphinx-prompt 
 sphinx_substitution_extensions 
+sphinx-copybutton
 sphinxcontrib-mermaid
 pygments>=2.7.4 # not directly required, pinned by Snyk to avoid a vulnerability
 sphinx>=3.0.4 # not directly required, pinned by Snyk to avoid a vulnerability
