+++ openfl-workspace/torch_cnn_mnist/plan/plan.yaml
@@ -10,7 +10,7 @@ aggregator :
     last_state_path     : save/torch_cnn_mnist_last.pbuf
     rounds_to_train     : 10
     log_metric_callback :
-      template : src.mnist_utils.write_metric
+      template : src.utils.write_metric
 
 
 collaborator :
@@ -22,7 +22,7 @@ collaborator :
 
 data_loader :
   defaults : plan/defaults/data_loader.yaml
-  template : src.ptmnist_inmemory.PyTorchMNISTInMemory
+  template : src.dataloader.PyTorchMNISTInMemory
   settings :
     collaborator_count : 2
     data_group_name    : mnist
@@ -30,7 +30,7 @@ data_loader :
 
 task_runner :
   defaults : plan/defaults/task_runner.yaml
-  template : src.pt_cnn.PyTorchCNN
+  template : src.taskrunner.PyTorchCNN
 
 network :
   defaults : plan/defaults/network.yaml
