+++ openfl-workspace/workspace/plan/defaults/tasks_torch.yaml
@@ -1,19 +1,19 @@
 aggregated_model_validation:
-  function : validate
+  function : validate_task
   kwargs   :
     apply   : global
     metrics :
       - acc
   
 locally_tuned_model_validation:
-  function  : validate
+  function  : validate_task
   kwargs    :
     apply: local
     metrics :
       - acc
   
 train:
-  function : train_batches
+  function : train_task
   kwargs   :
     metrics     :
     - loss
